[cascade.gateway] split spawning.py into subpackage
Move spawning logic into spawning/{__init__,common,wheels,local,ssh,slurm,troika}.py.
Private _-prefixed functions become module-public within their respective module.
spawn_subprocess dispatch and EkwInstallSpec/prepare_install_spec re-exports remain
in __init__.py so all existing callers are unchanged.
